requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.22.0,<1.0.0
gunicorn>=20.1.0,<21.0.0
google-cloud-vision>=3.4.2,<4.0.0
google-generativeai>=0.3.2,<1.0.0
google-cloud-aiplatform>=1.38.0,<2.0.0
pyyaml>=6.0,<7.0
